separated response into a class
